Support file time interpretation for '.time' files
This adds a second style of parsing, which inspects the file's unix
timestamp and interprets digits from the timestamp in a 2-digit sliding
window. For example, given a timestamp of "01080900":

"01" -> begin print str state
"08" -> add "h" char
"09" -> add "i" char
"00" -> end state and print

So this timestamp would print "hi".

Since the `touch` util doesn't support millisecond timestamps, multiple
".time" files are needed to perform more complex operations. As a
result, the interpreter was updated to also accept a directory of files,
which are sorted and interpreted in alphabetical order (see
examples/time/hello_world).