Begin refactoring from julia->racket
I haven't been terribly happy with julia's startup time, and gave up
pretty quickly on attempts to speed it up. The time it took to just open
and read a file was about 0.5s, which compared to other languages feels
pretty sluggish. I'm not sure if I was just doing something wrong, but
either way, I'm rewriting the interpreter in a language I've wanted to
try out for a while, Racket.

Functionally this new interpreter is behind the one written in Julia --
it can currently only interpret .space files -- but it functions a bit
better than the Julia one. The Julia one required developers to write the
number of spaces for the character they wanted + 1, which has been fixed
in the Racket implementation.