Update command mapping to allow storing variables
Variables can now be stored in Y2K programs, using a new 2-digit command
(04). Each variable accepts a name/digit, variable type, and size,
followed by the actual data to be stored in that variable.

Once a size has been established for the variable, the parser switches
to single digit parsing in order to reliably read in the value without
worrying about rollover between different files.

For example, to store the integer 100 in a variable "01", and then print
that variable, the timestamp sequence would be:

0401020310 -> next file -> 01000...

Breaking this down:
04 - begin set variable
01 - set var 01
02 - numeric type
03 - three digits
1  - digit 1
0  - digit 2
<new file>
01 - resume parsing
0  - digit 3
00 - end parsing
03 - begin print var
01 - set var to print to 01
00 - end parsing (print var)