Refactor variable struct to handle all data types
With a couple of small tweaks, the variable struct can actually handle
ints/floats/strings equally well, since everything is read in initially
as a string and then converted to another value as needed.

Some thoughts:
For floats, I think the order should go
<set><name><type><size><...data...><decimal pos>

So to write 3.14 it would be something like
04 01 03 03 314 01

With the trailing "01" meant to place the dot after the 3 ("00" would
indicate a leading 0, as in 0.314).

Likewise, for the eventual inclusion of basic arithmetic, I think it
could be possible to distinguish between variables and raw values using
a similar approach. Maybe something like

<add fn><num args>

then afterwards indicate for each arg whether it's a raw value or a
variable using a 99 for variables and 11 for regular values maybe (just
trying to keep the numbers a little easier to read perhaps). So to
process "var(01) + 5" it would look something like

<add fn> 03 99 11 01 01 05

There are a lot of 0Xs in the examples so far, so I'd like to start
including different number combinations to keep it interesting and
marginally easier to read at first glance.