Rewrite interpreter from Racket to Scheme
After learning a bit more about Racket and Scheme in general, I decided
to give Scheme a try (specifically Chicken Scheme, due to its ability to
compile to C code, which could help with portability of the interpreter
later on), and I ended up liking the relative simplicity of it compared
to Racket. I'm not really sure which implementation I'll move forward
yet, but it was pretty quick to rewrite the existing Racket code in
Scheme either way.