Refactor most alists to generic lists
This introduces a small readability issue, where lists are accessed via
an index provided in the timestamp, but it isn't as clear when reading
through the code that that's what's happening. Will need to reassess
whether or not this is the right approach. I sense a more major refactor
in the works to try and finally make this readable for anyone except for
me.