Remove adjustments to parsing size
Previously the parsing window would shrink from 2-digits to 1-digit when
parsing a variable in order to accurately store numbers whose total
digits was an odd number (100 for example).

This instead enforces a constant 2-digit parsing window to avoid any
confusion when creating a program. Keeping track of where the odd digit
ended ruined the ability to quickly discern a program's goal. Instead,
for odd-digit numbers, the remaining digit is ignored.

For example, to set variable 01 to 100, we would have:

04 01 02 03 10
10 00 .. .. ..

In the second file, the variable creator will see "00" but will only
pull the first "0" into the new variable and ignore the second one. The
second file could also have:

10 09 .. .. ..

and the outcome would be the same, since the 9 will be ignored.