Rewrite variable creation into conditional sequence
The flow of creating a variable previously seemed hard to follow, with a
set of lambdas being accessed based on a globally modified index. Not
super clear.

This new approach structures variable creation around a single function
that conditionally checks in order for which variable elements have been
set. The parsing loop continues to pass values to this function until
all fields have been finalized, at which point the parsing mode is set
back to "standby".