Rewrite interpreter in Go
I'm tired of forcing myself to try and like LISP syntax. The number of
parentheses is just kinda ugly imo. This was just supposed to be a fun
project, but it feels like work the longer that I spend trying to do
something I don't like. I should probably start with something smaller
if I want to try learning to enjoy LISP syntax.

Rewriting in Go to be more legible (for me) and to improve the overall
quality of code since it's a language I already know (and don't have to
spend extra time stumbling through learning a new language).