Use microsecond precision in examples
Using microsecond precision allows "programs" to be quite a bit smaller,
especially when combined with single digit parsing. Setting and printing
a variable, for example, is now only a single file, and it doesn't use
all available digits. Setting, modifying, and printing a variable can be
done in a single file as well, using all but one digit.