Allow modifying interpreter state via meta command
Interpreter state (debug and digit count) can now be modified within
programs. This will be especially helpful for new users, since they
won't have to remember argument values to run the different examples.

The only example that needed to be updated was the hello world one,
since that has to use a 2-digit parsing window. For this example, a meta
command was added to the beginning of the program to switch to a 2-digit
window from a 1-digit window (1-digit parsing is the new default).