Add n-term fib example, catch unset var errs
Added an example that uses command line arguments to specify a number of
fibonacci terms to print.

Accessing variables that have not yet been set is now protected against,
and will return an empty var struct when attempted. This allows programs
to not crash immediately, but will still alert the user that they are
likely referencing something incorrectly.