Remove `-raw` flag
The "-raw" flag wasn't very "ergonomic" for newcomers to the language,
especially since all examples required that flag in order to work.

As a result, the "-raw" flag is no longer required to run "raw" files.
Instead, the interpreter evaluates in the following order:

Is the input a directory? If not, was it modified before 2000? If not,
assume that it's a raw Y2K file.

This allows new users to run Y2K in the following ways, without any
differences or errors:

(Raw file)
$ y2k examples/set-and-print-var.y2k

(Exported directory)
$ y2k ./y2k-out

(Single exported timestamp file)
$ y2k ./y2k-out/0.y2k