Insert variables when referenced before creation
This change automatically adds new variables whenever they are
referenced before creation. So if a variable is referenced in a MODIFY
or CONDITION command, it will be inserted as an empty int variable, with
a value of 0.

This is purely done for code-golfing reasons, as this can mean the
difference between 1 file and 2+.