Support `float` variable type
Float variables can now be created in Y2K programs, but involve a
slightly different approach than other variable types.

To create a float variable, the specified size should be 1 digit longer
than desired, because the very first digit is used for decimal
placement. For example, the number 3.14 would be written as:

81341314

In the above example, we create variable 1 with type float (3) and size
4, which inserts the values 1314. Because we specified float type
though, the first digit is removed and inserted after the new first (1)
digit, which is 3. The result is a variable with the value 3.14.