Add "isVariable" field to `MODIFY` cmd
An additional field has been added after the MODIFY command's "Function"
field has been set. This is a boolean field to indicate whether the
function argument should be treated as a primitive value that matches
the modifying target variable's data type, or if it should be treated as
a variable ID.

For example, 711011 would call "var 1 += 1". 711112 would call "var 1 +=
var 2". The third field after the command ID is seen to switch from 0 to
1 in between these two examples, which switches the interpreter between
interpreting the value as a regular primitive type versus a variable.

Examples have been updated, as well as the documentation.