Remove unnecessary switch statements
A lot of the switch statements used in the interpreter only had a single
condition and a default case. This is because variables are only ever
treated as either a string or a numeric value. With this in mind, it
makes more sense to just have an `if` check to see if the variable is a
string, otherwise assume that the variable's numeric value is being
updated.